chore: refactor label subcomponent
