--- packages/react/src/components/_helpers/Label/index.ts
@@ -1,4 +1,0 @@
-import Label, { LabelProps } from './Label'
-
-export { Label, LabelProps }
-export default Label
