+++ packages/react/src/components/_subcomponents/Label/Label.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import renderWithTheme from '../../../helpers/renderWithTheme'
-import Label from './Label'
+import Label from './Label.subcomponent'
 
 describe('Label', () => {
   it('should render correctly', () => {
@@ -11,6 +11,18 @@ describe('Label', () => {
 
   it('should render correctly when is disabled', () => {
     const { styles, component } = renderWithTheme(<Label text="some text" disabled />)
+
+    expect([styles, component.container]).toMatchSnapshot()
+  })
+
+  it('should render correctly with the given color', () => {
+    const { styles, component } = renderWithTheme(<Label text="some text" color="mediumEmphasis" />)
+
+    expect([styles, component.container]).toMatchSnapshot()
+  })
+
+  it('should render correctly with the given fontSize', () => {
+    const { styles, component } = renderWithTheme(<Label text="some text" fontSize={18} />)
 
     expect([styles, component.container]).toMatchSnapshot()
   })
