+++ packages/react/src/components/_subcomponents/Label/index.ts
@@ -1,0 +1,4 @@
+import Label, { LabelSubcomponentProps } from './Label.subcomponent'
+
+export { Label, LabelSubcomponentProps }
+export default Label
